// src/tools/model/SignupModel.tsx
export type SignupModel = {
  userId: string;
  userPw: string;
  userName: string;
  residentNumber: string;
  gender: string;
  engName: string;
  task: string | undefined;
  empNo: string;
  hiredate: Date | string;
  family: string;
  armyStart: Date | string;
  armyEnd: Date | string;
  armyBranch: string;
  birthday: Date | string;
  tel: string;
  emergencyTel: string;
  userEmail: string;
  address: string;
  techGrade: string | undefined;
  position: string | undefined;
  spot: string | undefined;
  loc: string | undefined;
  locDetail: string | undefined;
  scienceTechCertify: string;
  admincheck: number;
  highschool: string;
  education: string;
  gradSchool: string;
};